Added fixes for Airflow to be usable on Windows Dask-Workers (#30249)

+++ airflow/utils/process_utils.py
@@ -80,6 +80,8 @@ def reap_process_group(
         returncodes[p.pid] = p.returncode
 
     def signal_procs(sig):
+        if IS_WINDOWS:
+            return
         try:
             logger.info("Sending the signal %s to group %s", sig, process_group_id)
             os.killpg(process_group_id, sig)
@@ -108,7 +110,7 @@ def reap_process_group(
             else:
                 raise
 
-    if process_group_id == os.getpgid(0):
+    if not IS_WINDOWS and process_group_id == os.getpgid(0):
         raise RuntimeError("I refuse to kill myself")
 
     try:
